Remove types file -- simplify structure
