+++ src/index.ts
@@ -1,4 +1,4 @@
 import { RNGeth } from "./RNGeth";
-export { NodeConfig} from "./types";
+export { NodeConfig} from "./RNGeth";
 
 export default RNGeth
